Moves Exceptions handler and response to shared folder
